I combined the 2 files together
